requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-SocketIO>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.7.0
Werkzeug>=2.3.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
requests>=2.31.0
docker>=6.1.0
websocket-client>=1.6.0
gunicorn>=21.2.0
eventlet>=0.33.3
psycopg2-binary>=2.9.6

